Added name based arguments
Added 3 additional commands:

Update | Updates the offline "data.txt" to include the full list of files and their ID parts as well as create the file if it does not exist. All information is generated and changed over every "update". The mode is changed to allow "silent" updates of "data.txt"

Grab | Allows a user to grab a file based on name. This lets users download files without needing to type or copy paste an ID. It is ideal for mobile terminals and touchpad users. It is limited by name, however, as files with the same name will change the ID value in "data.txt", meaning a user has to fall back to UDS "pull" methods.

Erase | Allows a user to delete a file based on name. It follows the same limitations as "Grab", meaning files with same name will cause problems.

Removed a few values that were unused, as their purpose is yet to be known...